cmd/compile: add likely, unlikely keywords
This keywords are effectively compiler instrinsics that allow manual setting of
the likely, unlikely directions for code generation. These bits are
automatically proagated through simple expressions, such as && and ||.

The use of these instrinics should be minimal, but they can have a significant
effect on performance, particularly for functions in the hot path that follow
Go style guidelines. For example, the following set of benchmark illustrates
this point.

	package main

	import (
		"fmt"
		"testing"
	)

	var enabled bool = false

	//go:noinline
	func LogIdiomatic(message string) {
		if !enabled {
			return
		}
		fmt.Printf("expensive, never done\n")
	}

	//go:noinline
	func LogNonStandard(message string) {
		if enabled {
			fmt.Printf("expensive, never done\n")
		}
	}

	//go:noinline
	func LogLikely(message string) {
		if likely(!enabled) {
			return
		}
		fmt.Printf("expensive, never done\n")
	}

	func BenchmarkLogIdiomatic(b *testing.B) {
		s := "message"
		for i := 0; i < b.N; i++ {
			LogIdiomatic(s)
		}
	}

	func BenchmarkLogNonStandard(b *testing.B) {
		s := "message"
		for i := 0; i < b.N; i++ {
			Log2(s)
		}
	}

	func BenchmarkLogLikely(b *testing.B) {
		s := "message"
		for i := 0; i < b.N; i++ {
			LogLikely(s)
		}
	}

	func TestPreventInlining(t *testing.T) {
		s := "message"
		enabled = true
		LogIdiomatic(s)
		LogNonStandard(s)
		LogLikely(s)
		enabled = false
	}

Above, LogIdiomatic represents the recommended idiomatic style in Go, and yet
produces the worst results. This is because the expression is treated as
unlikely, though it should not be. The compiler is unaware of this fact, and
the programmer must provide an appropriate hint.

The LogLikely variant uses the idiomatic form, but an appropriate hint is
provided that allows the compiler to produce more optimal code for this
function. The results are as follows:

	goos: linux
	goarch: amd64
	BenchmarkLogIdiomatic-4        	10000000000	         2.34 ns/op
	BenchmarkLogNonStandard-4       10000000000	         2.02 ns/op
	BenchmarkLogLikely-4   		10000000000	         2.01 ns/op

Note that the likely and unlikely functions appear as built-ins, but can be
shadowed by other expressions, so existing code should not break.